refactor: constants 파일 utils/로 이동
