+++ src/helper/insertMoneyHelper.js
@@ -1,4 +1,4 @@
-import constants from "mockData/constants";
+import constants from "utils/constants";
 import numberUtil from "utils/numberUtil";
 
 const { seperateThousands } = numberUtil;
